laser_detection_model/requirements.txt
pyrealsense2
numpy
Pillow
ultralytics
dvc
dvc[s3]
python-dotenv
labelbox
labelbox[data]
ndjson
natsort
../../laser_control
